Set a default username/password
Set a default username of `php-thrift-sql` when there is no username given for either Hive or Impala.

Also set a default password for Hive if none given. With SASL Hive just needs something set to not error out which prevents some common problems when using this lib. This closes #22 and closes #23.