Swap out classmap generator
Start using `theseer/autoload` instead of `jesseschalken/autoload-generator` to generate classmap used for autoloading the PHAR. This is done because the jesseschalken version is quite old now, last update 2016, and is using `nikic/php-parser` 1.4 which has a bug during tokenization. That bug causes `echo()` method definitions to be confused with `echo()` function redefinitions forcing us to use a monkey patch to change the Thrift Beeswax interface to avoid using `BeeswaxServiceIf::echo()`.

By switching out to the newer `theseer/autoload` classmap generator, last update May 2019, we get to use PHP's "native" `ext/tokenizer` extension instead which does a better job tokenizing and does not have a problem with `echo()` methods allowing us to get rid of the hacky monkey patch and follow the Thrift service defination more closely. The only downside of this new classmap generator does not force preloading of `Thrift/Transport/TSocketPool.php` which the classmap generator did due to the use of `function_exists('apc_fetch')` in that file. In theory this lazyloading should not matter in our particular case but may change if `TSocketPool` starts trying to do something fancier then a noop ployfill of APC.